Move Page class out of its own file
