+++ iOS-sample-app/Showcases/Screen Views/Paged Scroll Views/PagedScrollViewController.swift
@@ -72,3 +72,19 @@ class PagedScrollViewController: UIViewController, UIScrollViewDelegate {
         layout(withSize: size)
     }
 }
+
+class Page: UIView {
+    static func instanceFromNib(index: Int) -> Page {
+        let page = Bundle.main.loadNibNamed("Page", owner: self, options: nil)?.first as! Page
+        page.index = index
+        return page
+    }
+    
+    @IBOutlet weak var pageNumberLabel: UILabel!
+
+    var index = 1 {
+        didSet {
+            pageNumberLabel.text = String(index)
+        }
+    }
+}
